Refactor settings route into focused panels
